Вынесет общую логику обновления
